requirements.txt
gradio==4.44.1
google-generativeai==0.3.2
Pillow==10.3.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.31.0
numpy==1.24.3
matplotlib==3.7.2 